rename category mapper to ICategoryMapper to make it clear it's an interface
