+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/category/ListCategoriesGatewayImpl.java
@@ -2,7 +2,7 @@ package br.com.jpcchaves.infrastructure.gatewayimpl.category;
 
 import br.com.jpcchaves.application.category.gateway.ListCategoriesGateway;
 import br.com.jpcchaves.core.domain.Category;
-import br.com.jpcchaves.infrastructure.mapper.CategoryMapper;
+import br.com.jpcchaves.infrastructure.mapper.contracts.ICategoryMapper;
 import br.com.jpcchaves.infrastructure.persistence.entity.CategoryEntity;
 import br.com.jpcchaves.infrastructure.persistence.repository.ICategoryRepository;
 import java.util.List;
@@ -14,10 +14,10 @@ public class ListCategoriesGatewayImpl implements ListCategoriesGateway {
   @Qualifier("${deploy.repo}")
   private final ICategoryRepository<CategoryEntity, Long> repository;
 
-  private final CategoryMapper categoryMapper;
+  private final ICategoryMapper categoryMapper;
 
   public ListCategoriesGatewayImpl(
-      ICategoryRepository<CategoryEntity, Long> repository, CategoryMapper categoryMapper) {
+      ICategoryRepository<CategoryEntity, Long> repository, ICategoryMapper categoryMapper) {
     this.repository = repository;
     this.categoryMapper = categoryMapper;
   }
