+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/category/UpdateCategoryGatewayImpl.java
@@ -4,7 +4,7 @@ import br.com.jpcchaves.application.category.gateway.UpdateCategoryGateway;
 import br.com.jpcchaves.core.domain.Category;
 import br.com.jpcchaves.core.exception.CategoryException;
 import br.com.jpcchaves.core.exception.enums.ExceptionDefinition;
-import br.com.jpcchaves.infrastructure.mapper.CategoryMapper;
+import br.com.jpcchaves.infrastructure.mapper.contracts.ICategoryMapper;
 import br.com.jpcchaves.infrastructure.persistence.entity.CategoryEntity;
 import br.com.jpcchaves.infrastructure.persistence.repository.ICategoryRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -15,10 +15,10 @@ public class UpdateCategoryGatewayImpl implements UpdateCategoryGateway {
   @Qualifier("${deploy.repo}")
   private final ICategoryRepository<CategoryEntity, Long> repository;
 
-  private final CategoryMapper categoryMapper;
+  private final ICategoryMapper categoryMapper;
 
   public UpdateCategoryGatewayImpl(
-      ICategoryRepository<CategoryEntity, Long> repository, CategoryMapper categoryMapper) {
+      ICategoryRepository<CategoryEntity, Long> repository, ICategoryMapper categoryMapper) {
     this.repository = repository;
     this.categoryMapper = categoryMapper;
   }
